requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic[email]
pytest
black
pytest
pytest-cov
bcrypt
python-multipart
python-dotenv